fixup! Added stretch and offset prop to Popover
